AttemptEntryIntoBuilding V0.01 - Stable
This version is stable and works. If anything fails, revert to this.
This works. NPCs will check for an active window that isn't blocked off when they check a door. (95% of the time) vs how it was previously (where the NPC wouldn't check anything.)

This was tested only for the npc being outside.